There were some issues with the ApiCall component. The API was already returning an array, but inside the setApiData function, you were passing data inside another array (setApiData([json])). This resulted in no data being available to display on the page. Therefore, I removed the extra array in the setApiData function. Additionally, the file was a JavaScript (.js) file, which was causing a compiler error. I changed the file extension to JSX (.jsx), and now everything is working correctly.
