+++ src/App.jsx
@@ -1,11 +1,12 @@
 import { useState } from 'react'
 import './App.css'
+import ApiCall from './components/ApiCall'
 
 function App() {
 
   return (
     <>
-     
+     <ApiCall />
     </>
   )
 }
